Move Impl classes to own files
