+++ core/identity-spi/src/main/java/software/amazon/awssdk/identity/spi/AwsCredentialsIdentity.java
@@ -15,12 +15,7 @@
 
 package software.amazon.awssdk.identity.spi;
 
-import java.util.Objects;
-import software.amazon.awssdk.annotations.Immutable;
-import software.amazon.awssdk.annotations.SdkInternalApi;
 import software.amazon.awssdk.annotations.SdkPublicApi;
-import software.amazon.awssdk.utils.ToString;
-import software.amazon.awssdk.utils.Validate;
 
 /**
  * Provides access to the AWS credentials used for accessing services: AWS access key ID and secret access key. These
@@ -53,116 +48,4 @@ public interface AwsCredentialsIdentity extends Identity {
      * Retrieve the AWS secret access key, used to authenticate the user interacting with services.
      */
     String secretAccessKey();
-}
-
-@Immutable
-@SdkInternalApi
-final class AwsCredentialsIdentityImpl implements AwsCredentialsIdentity {
-
-    private final String accessKeyId;
-    private final String secretAccessKey;
-
-    AwsCredentialsIdentityImpl(String accessKeyId, String secretAccessKey) {
-        this.accessKeyId = Validate.paramNotNull(accessKeyId, "accessKeyId");
-        this.secretAccessKey = Validate.paramNotNull(secretAccessKey, "secretAccessKey");
-    }
-
-    @Override
-    public String accessKeyId() {
-        return accessKeyId;
-    }
-
-    @Override
-    public String secretAccessKey() {
-        return secretAccessKey;
-    }
-
-    @Override
-    public String toString() {
-        return ToString.builder("AwsCredentialsIdentityImpl")
-                       .add("accessKeyId", accessKeyId)
-                       .build();
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        AwsCredentialsIdentityImpl that = (AwsCredentialsIdentityImpl) o;
-        return Objects.equals(accessKeyId, that.accessKeyId) &&
-               Objects.equals(secretAccessKey, that.secretAccessKey);
-    }
-
-    @Override
-    public int hashCode() {
-        int hashCode = 1;
-        hashCode = 31 * hashCode + Objects.hashCode(accessKeyId());
-        hashCode = 31 * hashCode + Objects.hashCode(secretAccessKey());
-        return hashCode;
-    }
-}
-
-@Immutable
-@SdkInternalApi
-final class AwsSessionCredentialsIdentityImpl implements AwsSessionCredentialsIdentity {
-
-    private final String accessKeyId;
-    private final String secretAccessKey;
-    private final String sessionToken;
-
-    AwsSessionCredentialsIdentityImpl(String accessKeyId, String secretAccessKey, String sessionToken) {
-        this.accessKeyId = Validate.paramNotNull(accessKeyId, "accessKeyId");
-        this.secretAccessKey = Validate.paramNotNull(secretAccessKey, "secretAccessKey");
-        this.sessionToken = Validate.paramNotNull(sessionToken, "sessionToken");
-    }
-
-    @Override
-    public String accessKeyId() {
-        return accessKeyId;
-    }
-
-    @Override
-    public String secretAccessKey() {
-        return secretAccessKey;
-    }
-
-    @Override
-    public String sessionToken() {
-        return sessionToken;
-    }
-
-    @Override
-    public String toString() {
-        return ToString.builder("AwsSessionCredentialsIdentityImpl")
-                       .add("accessKeyId", accessKeyId())
-                       .build();
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-
-        AwsSessionCredentialsIdentityImpl that = (AwsSessionCredentialsIdentityImpl) o;
-        return Objects.equals(accessKeyId, that.accessKeyId) &&
-               Objects.equals(secretAccessKey, that.secretAccessKey) &&
-               Objects.equals(sessionToken, that.sessionToken);
-    }
-
-    @Override
-    public int hashCode() {
-        int hashCode = 1;
-        hashCode = 31 * hashCode + Objects.hashCode(accessKeyId());
-        hashCode = 31 * hashCode + Objects.hashCode(secretAccessKey());
-        hashCode = 31 * hashCode + Objects.hashCode(sessionToken());
-        return hashCode;
-    }
 }
